Disabling SF-SAC download button when unavailable (#4139)
* Handling disabling SF-SAC download button

* Lint

* Using tooltip instead of title

* Fixing button styling

* Lint

* Fixing unit tests

* Lint

* Lint

* Adding unit test

* Lint

* Lint

* Adding noreferrer

* Using real FAQ link
