+++ backend/dissemination/test_search.py
@@ -1,5 +1,3 @@
-import os
-from django.db import connection
 from django.test import TestCase
 from dissemination.models import (
     DisseminationCombined,
@@ -15,6 +13,7 @@ from dissemination.search import (
     search,
     is_advanced_search,
 )
+from dissemination.test_materialized_view_builder import TestMaterializedViewBuilder
 
 from model_bakery import baker
 
@@ -451,31 +450,6 @@ class SearchGeneralTests(TestCase):
         # Searching for several IDs yields the same number of results
         results = search_general(General, {"report_id": report_ids[2:4]})
         self.assertEqual(len(results), 2)
-
-
-class TestMaterializedViewBuilder(TestCase):
-    def setUp(self):
-        super().setUp()
-        self.execute_sql_file("dissemination/sql/create_materialized_views.sql")
-
-    def tearDown(self):
-        self.execute_sql_file("dissemination/sql/drop_materialized_views.sql")
-        super().tearDown()
-
-    def execute_sql_file(self, relative_path):
-        """Execute the SQL commands in the file at the given path."""
-        full_path = os.path.join(os.getcwd(), relative_path)
-        try:
-            with open(full_path, "r") as file:
-                sql_commands = file.read()
-            with connection.cursor() as cursor:
-                cursor.execute(sql_commands)
-        except Exception as e:
-            print(f"Error executing SQL command: {e}")
-
-    def refresh_materialized_view(self):
-        """Refresh the materialized view"""
-        self.execute_sql_file("dissemination/sql/refresh_materialized_views.sql")
 
 
 class SearchALNTests(TestMaterializedViewBuilder):
